Calculate the start time of a stream differently.
For reasons that I don't understand, the first packet or two in an ffmpeg
capture (at least on OpenBSD) are borderline nonsense (e.g. they last too long
and have dodgy data in them). They also seem to have inaccurate wall-clock times
in them. A bit of fiddling suggests that the least unreliable mechanism for
guessing the real start time of the stream is to use an offset from the time of
the first packet. This has worked for a couple of videos I've tried, but whether
it holds up over time remains to be seen!