Force whole-frame latencies by rounding down to the nearest frame.
I found this out the hard way: if your latency calculation is intra-frame, you
can get some entertaining results with the "fps" filter in ffmpeg. In
particular, imagine you have a framerate of 30fps, but your latency calculation
happens to be equivalent to 60fps: it ends up being almost random as to whether
any given video frame should be rounded up or down to align to a frame boundary.
This leads to frame dropping and duplication. There are certainly good reasons
for wanting latencies that are finer-grained than a frame, but I'd rather save
myself the future confusion. This also means that moving to "nearest" rounding
in fps makes sense (again).