Perform more mixed scene processing in one pass.
Unfortunately, some ffmpeg filters cause it to buffer arbitrary quantities of
data (the overlay filter appears to be a particular, but not the only, culprit).
Previously we thus created each complicated scene independently, causing us to
have to do multiple passes over the source files, which, for long inputs, is
incredibly slow (processing taking multiple hours). This commit breaks this up
significantly: we now do many fewer passes over the source files (we still
possibly do more than one pass, because ffmpeg can still gobble up GiB of memory
if you're not careful, but we process 10 files per pass). This does mean that we
have to create more intermediate files, but this is more than worth it.